427: Make files upload to the folder you are in (#472)
There was a bug where the file upload form was not aware of your folder
traversal.

This is resolved by instead re-rendering the file explorer on a folder
navigation.

The endpoint takes 800ms to render when I do this locally, but we need
some kind of performance-based epic to go through code and determine how
to speed everything up later.

---------

Signed-off-by: James Orson <jamesaorson@gmail.com>